requirements.txt
python-dotenv
openai
aiogram
transformers